Remove CheckoutCoordinator ViewModel.
